Bug 1926291 - don't kill pipewire in interactive tasks. r=jmaher
During interactive tasks, test-linux.sh exits without running the tests,
so we should keep the session running.

Differential Revision: https://phabricator.services.mozilla.com/D229128

+++ taskcluster/scripts/tester/test-linux.sh
@@ -76,7 +76,7 @@ maybe_start_pulse() {
     fi
 }
 cleanup_pipewire() {
-    if [ -n "$pw_pids" ]; then
+    if [ -n "$pw_pids" ] && [ $TASKCLUSTER_INTERACTIVE = false ]; then
         kill "${pw_pids[@]}"
     fi
 }
